Move authentication wrapper to middleware.
